fix: enforce little-endian byte order for binary PCM WebSocket frames
numpy.tobytes() uses native endianness which may not match the
frontend's Float32Array on non-little-endian platforms. Now explicitly
encodes as little-endian float32 via astype('<f4').

Fixes #2875

+++ auralis-web/backend/core/audio_stream_controller.py
@@ -1257,7 +1257,7 @@ class AudioStreamController:
                             "stream_type": stream_type,
                         },
                     }
-                    pcm_bytes: bytes = frame_samples.tobytes()
+                    pcm_bytes: bytes = frame_samples.astype('<f4').tobytes()
                     await queue.put((metadata, pcm_bytes))
             finally:
                 await queue.put(None)  # sentinel
